requirements.txt
git+https://github.com/pyrogram/pyrogram@master
python-dotenv
pykeyboard
pydantic
tgcrypto
pyromod
